Use ISO 8601 timestamp format in JSON to Loggly
The Loggly documentation states that the timestamp field must be in ISO 8601 format for Loggly to recognize it. This way the actual time displayed in Loggly will be based on the value sent. Currently when sending the UNIX time, Loggly will not recognize it and will display the time it received the message instead.

https://www.loggly.com/docs/automated-parsing/#json